+++ src/pystackapi/__init__.py
@@ -1,5 +1,5 @@
 from .network import Network
 from .site import Site
 
-__version__ = '0.2.0dev'
+__version__ = '0.2.0'
 __all__ = ('Site', 'Network', 'sites', 'errors', 'item')