+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Added:
 
+ + method `add_answer` to class `Site`;
  + method `get_me` to class `Site`;
  + method `get_my_inbox` to class `Site`;
+ + method `get_my_full_reputation_history` to class `Site`;
  + method `get_my_unread_inbox` to class `Site`;
  + method `get_unanswered_questions_on_my_tags` to class `Site`;
  + method `post` to class `Site`